Removed extra content folder, work on enemies
Using a "knoll/gnoll" free asset as an enemy, redid a lot of the
blueprinting to get it working in my project. It can currently follow
the player around while in a navmesh volume and performs an attack
animation when it gets close. Also found I had a duplicate of some of
the standard assets (probably all of them honestly) from when I was
transferring them into the project. It is now gone so the repo is about
half a gig smaller.