Work on dungeon and destructible objects
Got my first destructible object working, a rock that shatters into
pieces when you throw it hard enough. Updated Grabber to set up the
FCollisionObjectQueryParams object outside of the trace, allowing me to
add destructibles to the object type list so you can throw both types
(destructibles and normal physics bodies for onces we don't want to
shatter).

Worked a lot more on the dungeon, mostly got rid of the huge slab floor
that made up most of the dungeon, was too limiting in making
multi-leveled more interesting dungeon layouts. Main room has 4 exits
(and no floor), hope to put a bit of a puzzle in each one that needs to
be completed to open a door at the far side of the main room.

Worked even more on materials a bit, didn't actually do much but learned
more about material blending types and the type of nodes that are
available in the material editor. Would maybe like to learn how to get
noise working with texcoords to make a noise based texture overlay.