Fixed fireball collisions, made spiral stairs in entryway linear
The fireball wasn't properly colliding with stuff, and when it was
enabled it wasn't knocking physics bodies/rocks around. It now hits
stuff and dies after a delay (so the particle system can properly die)
and knocks the rocks around as well, and can destroy the destructible
ones.

Made the spiral staircases in the entryway linear ones. It looks sort of
less interesting but the spiral ones were BSPs so there was a lot of
annoying behaviors, and it was hard to get up them. Tried to rotate them
and stuff to make it better but the only really good way would've been
shrinking the character itself, which probably isn't worth it for
getting up some spiral staircases.