Fix torch light culling, fix texture on destructible rock shards, change water material color
Took the lighting on the torches off of the particle and just added a
point light to the torches instead. Realized it was culling the light so
badly because it was attached to the particle system, which isn't
rendered off-screen. I thought it was the entire object being culled.
Joe will be doing ambient light via a big directional soon.

Found a little setting in the chunk parameters of the destructible rock,
which seems to have fixed the inner polygons not having the rock
texture.

Water is now more blue, which in game looks more glowy green, like sea
water. Looks kinda nice. Still not translucent.