Enemies can damage player. Slime death animation. Cleaned up a few things.
Both enemies can damage the players, slimes using their damage variable
(based on the amount of red in their color), the knoll damaging based on
how far the player is from the club when it lands, doing damage in a
radius around it of about 1.5 meters. Could probably used to be scaled
up or something, he's easy to avoid (but can probably 2-shot you with
perfect hits). Also did a sound effect when the club lands. I think
currently he's undamageable? I was having issues, may be due to the
changes to the fireball to make it damage both enemies. Will mess with
soon.