Finished arena puzzle
Slimes and knolls both now have an Activates Next that works the same as
the one on the activatables. Arena door is closed via pressure plate
once you enter the room, and is reopened by the gem. The gem is hidden
behind a door that is opened once all of the slimes are killed. Each
slime activates the door, and the door queries a trigger volume
containing the room for the # of slimes left within it. If it's 0, it
opens and destroys the trigger volume (because it was intercepting the
activation raytrace). To aid in this, activatables now have Destroys
Self On Activate and Locks Self On Activate properties, which are self
explanatory and only take effect after it has called its Activates
Nexts.